fix: make fixes

+++ src/api.js
@@ -21,7 +21,7 @@ const onSuccess = (response) => {
 const onFail = (err) => {
   const { response } = err
 
-  if (response.status === Error.UNAUTHORIZED) {
+  if (response && response.status === Error.UNAUTHORIZED) {
     localStorage.clear()
     window.location.reload()
     throw err
